check only admin can access

+++ app/Http/Controllers/AdminCategoryProductController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use App\Models\CategoryProduct;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class AdminCategoryProductController extends Controller
 {
@@ -16,8 +17,12 @@ class AdminCategoryProductController extends Controller
      */
     public function index()
     {
-        $categoryProducts = CategoryProduct::all();
-        return view('admin.Category_Products.index')->with('categoryProducts', $categoryProducts);
+        if(Auth::guard('admin')){  
+            $categoryProducts = CategoryProduct::all();
+            return view('admin.Category_Products.index')->with('categoryProducts', $categoryProducts);
+        } else {
+            return redirect()->route('admin.login')->with('status','Logout Sucessfuy');//back to login pages and control
+        }
     }
 
     /**
@@ -27,13 +32,17 @@ class AdminCategoryProductController extends Controller
      */
     public function create()
     {
-        $categories = Category::all();
-        $products = Product::all();
-
-        return view('admin.Category_Products.add')->with([
-            'categories' => $categories,
-            'products' => $products
-        ]);
+        if(Auth::guard('admin')){ 
+            $categories = Category::all();
+            $products = Product::all();
+    
+            return view('admin.Category_Products.add')->with([
+                'categories' => $categories,
+                'products' => $products
+            ]);
+        } else {
+            return redirect()->route('admin.login')->with('status','Logout Sucessfuy');//back to login pages and control
+        }
     }
 
     /**
@@ -44,17 +53,21 @@ class AdminCategoryProductController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'product_id' => 'required',
-            'category_id'=> 'required',
-        ]);
-
-        $categoryProduct = new CategoryProduct();
-        $categoryProduct->product_id = $request->product_id;
-        $categoryProduct->category_id = $request->category_id;
-        $categoryProduct->save();
-
-        return redirect()->route('admin.categoriesProducts')->with('success_message','your Category Products is create succesfully');
+        if(Auth::guard('admin')){
+            $request->validate([
+                'product_id' => 'required',
+                'category_id'=> 'required',
+            ]);
+    
+            $categoryProduct = new CategoryProduct();
+            $categoryProduct->product_id = $request->product_id;
+            $categoryProduct->category_id = $request->category_id;
+            $categoryProduct->save();
+    
+            return redirect()->route('admin.categoriesProducts')->with('success_message','your Category Products is create succesfully');
+        } else {
+            return redirect()->route('admin.login')->with('status','Logout Sucessfuy');//back to login pages and control
+        }
     }
 
     /**
@@ -65,8 +78,12 @@ class AdminCategoryProductController extends Controller
      */
     public function show($id)
     {
-        $categoryProduct =  CategoryProduct::find($id);
-        return view('admin.Category_Products.view')->with('categoryProduct', $categoryProduct);
+        if(Auth::guard('admin')){
+            $categoryProduct =  CategoryProduct::find($id);
+            return view('admin.Category_Products.view')->with('categoryProduct', $categoryProduct);
+        } else {
+            return redirect()->route('admin.login')->with('status','Logout Sucessfuy');//back to login pages and control
+        }
     }
 
     /**
@@ -77,16 +94,20 @@ class AdminCategoryProductController extends Controller
      */
     public function edit($id)
     {
-        $categoryProduct = CategoryProduct::where('id',$id)->first();
-        $categories = Category::all();
-        $products = Product::all();
-
-
-        return view('admin.Category_Products.edit')->with([
-            'categoryProduct' => $categoryProduct,
-            'categories' => $categories,
-            'products' => $products
-        ]);
+        if(Auth::guard('admin')){ 
+            $categoryProduct = CategoryProduct::where('id',$id)->first();
+            $categories = Category::all();
+            $products = Product::all();
+    
+    
+            return view('admin.Category_Products.edit')->with([
+                'categoryProduct' => $categoryProduct,
+                'categories' => $categories,
+                'products' => $products
+            ]);
+        } else {
+            return redirect()->route('admin.login')->with('status','Logout Sucessfuy');//back to login pages and control
+        }
 
     }
 
@@ -99,17 +120,21 @@ class AdminCategoryProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'product_id' => 'required',
-            'category_id'=> 'required',
-        ]);
-
-        $categoryProduct =  CategoryProduct::find($id);
-        $categoryProduct->product_id = $request->product_id;
-        $categoryProduct->category_id = $request->category_id;
-        $categoryProduct->save();
-
-        return redirect()->route('admin.categoriesProducts')->with('success_message','your Category Products is Update succesfully');
+        if(Auth::guard('admin')){
+            $request->validate([
+                'product_id' => 'required',
+                'category_id'=> 'required',
+            ]);
+    
+            $categoryProduct =  CategoryProduct::find($id);
+            $categoryProduct->product_id = $request->product_id;
+            $categoryProduct->category_id = $request->category_id;
+            $categoryProduct->save();
+    
+            return redirect()->route('admin.categoriesProducts')->with('success_message','your Category Products is Update succesfully');
+        } else {
+            return redirect()->route('admin.login')->with('status','Logout Sucessfuy');//back to login pages and control
+        }
     }
 
     /**
@@ -120,8 +145,12 @@ class AdminCategoryProductController extends Controller
      */
     public function destroy($id)
     {
-        $categoryProduct = CategoryProduct::find($id);
-        $categoryProduct->delete();
-        return redirect()->route('admin.categoriesProducts')->with('success_message','Your Category Products is Delete Successfully');
+        if(Auth::guard('admin')){ 
+            $categoryProduct = CategoryProduct::find($id);
+            $categoryProduct->delete();
+            return redirect()->route('admin.categoriesProducts')->with('success_message','Your Category Products is Delete Successfully');
+        } else {
+            return redirect()->route('admin.login')->with('status','Logout Sucessfuy');//back to login pages and control
+        }
     }
 }
